Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4639 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ gdiff2html/src/GDiff2HTML.java
@@ -32,7 +32,7 @@ public class GDiff2HTML
         {
             byte b = magic[i];
             out(lonib(b));
-            out(lonib(b>>4));
+            out(lonib((byte)(b>>4)));
         }
 		int vers = gdiff.read();
